refactor(frontend): consolidate composables into parent modules
- Merge useMonitorHistory into useMonitor
- Merge useNotificationSound into useNotification
- Merge useSelectedPath into useFileNavigation
- Remove empty useDraggable stub
