+++ frontend/src/App.vue
@@ -97,7 +97,7 @@ import { isTauri } from './composables/useTransport'
 import { isTouchDevice } from './composables/useTerminal'
 import { useI18n } from './composables/useI18n'
 import { isWebPreviewInput } from './utils/previewRouting'
-import { initMonitorHistory } from './composables/useMonitorHistory'
+import { initMonitorHistory } from './composables/useMonitor'
 import NotificationPanel from './components/notification/NotificationPanel.vue'
 import { useNotification } from './composables/useNotification'
 import { Settings, Bell, PanelRight, Plus, X, Star, AppWindow } from 'lucide-vue-next'
