+++ frontend/src/components/notification/NotificationCard.vue
@@ -21,7 +21,7 @@
 <script setup lang="ts">
 import { computed } from 'vue'
 import { X } from 'lucide-vue-next'
-import type { NotificationType } from '../../composables/useNotificationSound'
+import type { NotificationType } from '../../composables/useNotification'
 
 const props = defineProps<{
   type: NotificationType
