--- frontend/src/composables/useDraggable.ts
@@ -1,3 +1,0 @@
-export function useDraggable(_el: HTMLElement) {
-  // Drag functionality removed
-}
