--- frontend/src/composables/useSelectedPath.ts
@@ -1,7 +1,0 @@
-import { ref } from 'vue'
-
-const selectedPath = ref<string | null>(null)
-
-export function useSelectedPath() {
-  return { selectedPath }
-}
